Move shaping to its own class and file
